refactor(v0): Convert `v0` suffix into go.mod
Because we make no guarantee for stability and backwards compatibility,
migrate the suffix representing the unstable version into the Go module
definition.  It is common practice to set this here as opposed to using
a directory, but also since we aim to update this SDK frequently, the
directory itself would anyway be removed once we reach a stable v1.

This means that up until a 1.0 release of the SDK, users can import the
unstable `sdk.kraft.cloud/v0`.  Once we reach stability, we can change
the module name to `sdk.kraft.cloud` and beyond this we can use the `v2`
suffix.  This is recommended here[0].

[0]: https://go.dev/ref/mod#major-version-suffixes

Signed-off-by: Alexander Jung <alex@kraft.cloud>