refactor!: Rework the service request mechanism
This large refactor reworks how each request to KraftCloud's API is
performed by reworking the client into an individual request mechanism.
The new structure allows the user to customize a number of aspects of
the request without having to re-instantiate the client whilst
simultaneously making calls very readable.

The first idea presented in this change is introducing an abstract
`ServiceClient` which accepts a generic that represents the underlying
service interface that we wish to make calls against.  This
`ServiceClient` exposes options (including `WithMetro`, `WithTimeout`
and `WithHTTPClient`) which return the interface.  This enables
function chaining, for example:

```go
resp, _ := client.WithMetro("fra0").Create(...)
```

To make this chaining possible the rest client interface was broken up
into several parts including default set of options which the user sets
ahead of time and converting the `RESTClient` into a `ServiceRequest`
which handles individual request.  The structure adopts the default
options based on the initial construction but reads dynamically injected
values.

Each method in the instances service is reworked to adopt this change.

Signed-off-by: Alexander Jung <alex@kraft.cloud>