feat: Introduce `WithTimeout` and `WithRetries`
These have default values of 10 seconds and 3 retries and increases
the robustness of API calls on connections which timeout by allowing
for a retry _only when a timeout occurs_ and not when an error from
the API occurs.

Since this Go module is used by KraftKit, which can have long and
complex builds performing many requests via the SDK, this change
increases the likelihood of those invocations to succeed,
especially on flakey internet connections.

Signed-off-by: Alexander Jung <alex@unikraft.cloud>